Traffic Sign Recognition
This model recognises the traffic sign pictures. Dataset consisted of 39209 images with each of
size 30 x 30 x 3. Then I developed a CNN model using keras api. Trained the dataset on the
model with 15 epochs. Model gives 94.6% accuracy on testing data. I also developed a GUI
using tkinter where the user will select the image of the traffic sign and will pass it to the model.
The model will recognize the sign notation and will show output to the user.